getServices query (#36)
* Initial work in service creation

* Implement Sweet Alert2

* created the email confirmation modal

* Finished creation of email confirmation modal

* Implemented modal on userInfo

* added try...catch to http requests

* fixes

* implement query when fetching services

* fixed bug with populate

* remove console log
